PEARDEV-256 Q4 Chandan requirements
Pipeline needed to add correlation plots between protein and RNA.
In order to achieve this, protein data needed to be imported as
in the `AnnData` and `RDS` objects. Functions were changed for
this purpose in `results.py`. `Pydantic` data classes were
put in a separate file (`segment.py`). `extract_counts.py`
scripts was updated to take a second `XLSX` file as an argument
(the file which contains protein data).